move UI building to separate class
